refactor: move contract_reference_doc.py from shared/ to cli/
contract_reference_doc depends on adapters/udp and app/settings, which
violates the shared/ boundary (shared should not reach into adapters or
app). It is only consumed by the CLI tool and a test, so cli/ is the
correct home.

Closes #700
