+++ tools/config/generate_contract_reference_doc.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from vibesensor.shared.contract_reference_doc import render_contract_reference_markdown
+from vibesensor.cli.contract_reference_doc import render_contract_reference_markdown
 
 ROOT = Path(__file__).resolve().parents[2]
